Inject logger and relocate account DTOs
