+++ cs-project.Core/DTOs/LoginDTO.cs
@@ -1,0 +1,8 @@
+namespace cs_project.Core.DTOs
+{
+    public class LoginDTO
+    {
+        public required string UserName { get; set; }
+        public required string Password { get; set; }
+    }
+}
