+++ cs-project.Core/DTOs/RegisterDTO.cs
@@ -1,0 +1,9 @@
+namespace cs_project.Core.DTOs
+{
+    public class RegisterDTO
+    {
+        public required string UserName { get; set; }
+        public required string Email { get; set; }
+        public required string Password { get; set; }
+    }
+}
